+++ src/o365/spo/commands/orgnewssite/orgnewssite-remove.ts
@@ -1,0 +1,160 @@
+import auth from '../../SpoAuth';
+import { ContextInfo, ClientSvcResponse, ClientSvcResponseContents } from '../../spo';
+import config from '../../../../config';
+import request from '../../../../request';
+import commands from '../../commands';
+import GlobalOptions from '../../../../GlobalOptions';
+import {
+  CommandOption,
+  CommandValidate,
+  CommandError
+} from '../../../../Command';
+import SpoCommand from '../../SpoCommand';
+import Utils from '../../../../Utils';
+
+const vorpal: Vorpal = require('../../../../vorpal-init');
+
+interface CommandArgs {
+  options: Options;
+}
+
+interface Options extends GlobalOptions {
+  url: string;
+  confirm?: boolean;
+}
+
+class SpoOrgNewsSiteRemoveCommand extends SpoCommand {
+  public get name(): string {
+    return `${commands.ORGNEWSSITE_REMOVE}`;
+  }
+
+  public get description(): string {
+    return 'Removes a site from the list of organizational news sites';
+  }
+
+  public getTelemetryProperties(args: CommandArgs): any {
+    const telemetryProps: any = super.getTelemetryProperties(args);
+    telemetryProps.confirm = (!(!args.options.confirm)).toString();
+    return telemetryProps;
+  }
+
+  protected requiresTenantAdmin(): boolean {
+    return true;
+  }
+
+  public commandAction(cmd: CommandInstance, args: CommandArgs, cb: (err?: any) => void): void {
+    const removeOrgNewsSite = (): void => {
+      auth
+        .ensureAccessToken(auth.service.resource, cmd, this.debug)
+        .then((accessToken: string): Promise<ContextInfo> => {
+          if (this.debug) {
+            cmd.log(`Retrieved access token ${accessToken}. Retrieving request digest...`);
+          }
+
+          return this.getRequestDigest(cmd, this.debug);
+        })
+        .then((res: ContextInfo): Promise<string> => {
+          const requestOptions: any = {
+            url: `${auth.site.url}/_vti_bin/client.svc/ProcessQuery`,
+            headers: {
+              authorization: `Bearer ${auth.service.accessToken}`,
+              'X-RequestDigest': res.FormDigestValue
+            },
+            body: `<Request AddExpandoFieldTypeSuffix="true" SchemaVersion="15.0.0.0" LibraryVersion="16.0.0.0" ApplicationName="${config.applicationName}" xmlns="http://schemas.microsoft.com/sharepoint/clientquery/2009"><Actions><ObjectPath Id="64" ObjectPathId="63" /><Method Name="RemoveOrgNewsSite" Id="65" ObjectPathId="63"><Parameters><Parameter Type="String">${Utils.escapeXml(args.options.url)}</Parameter></Parameters></Method></Actions><ObjectPaths><Constructor Id="63" TypeId="{268004ae-ef6b-4e9b-8425-127220d84719}" /></ObjectPaths></Request>`
+          };
+
+          return request.post(requestOptions);
+        })
+        .then((res: string): void => {
+          const json: ClientSvcResponse = JSON.parse(res);
+          const response: ClientSvcResponseContents = json[0];
+          if (response.ErrorInfo) {
+            cb(new CommandError(response.ErrorInfo.ErrorMessage));
+            return;
+          }
+          else {
+            if (this.verbose) {
+              cmd.log(vorpal.chalk.green('DONE'));
+            }
+            cb();
+          }
+        }, (err: any): void => this.handleRejectedPromise(err, cmd, cb));
+    }
+
+    if (args.options.confirm) {
+      removeOrgNewsSite();
+    }
+    else {
+      cmd.prompt({
+        type: 'confirm',
+        name: 'continue',
+        default: false,
+        message: `Are you sure you want to remove ${args.options.url} from the list of organizational news sites?`,
+      }, (result: { continue: boolean }): void => {
+        if (!result.continue) {
+          cb();
+        }
+        else {
+          removeOrgNewsSite();
+        }
+      });
+    }
+  }
+
+  public options(): CommandOption[] {
+    const options: CommandOption[] = [
+      {
+        option: '-u, --url <url>',
+        description: 'Absolute URL of the site to remove'
+      },
+      {
+        option: '--confirm',
+        description: 'Don\'t prompt for confirmation'
+      }
+    ];
+
+    const parentOptions: CommandOption[] = super.options();
+    return options.concat(parentOptions);
+  }
+
+  public validate(): CommandValidate {
+    return (args: CommandArgs): boolean | string => {
+      if (!args.options.url) {
+        return 'Required parameter url missing';
+      }
+
+      const isValidSharePointUrl: boolean | string = SpoCommand.isValidSharePointUrl(args.options.url);
+      if (isValidSharePointUrl !== true) {
+        return isValidSharePointUrl;
+      }
+
+      return true;
+    };
+  }
+
+  public commandHelp(args: CommandArgs, log: (help: string) => void): void {
+    const chalk = vorpal.chalk;
+    log(vorpal.find(commands.ORGNEWSSITE_REMOVE).helpInformation());
+    log(
+      `  ${chalk.yellow('Important:')} before using this command, log in to a SharePoint Online tenant
+    admin site, using the ${chalk.blue(commands.LOGIN)} command.
+        
+  Remarks:
+
+    To remove a site from the list of organizational news sites, you have to
+    first log in to a tenant admin site using the ${chalk.blue(commands.LOGIN)} command,
+    eg. ${chalk.grey(`${config.delimiter} ${commands.LOGIN} https://contoso-admin.sharepoint.com`)}.
+
+  Examples:
+  
+    Remove a site from the list of organizational news sites
+      ${chalk.grey(config.delimiter)} ${commands.ORGNEWSSITE_REMOVE} --url https://contoso.sharepoint.com/sites/site1
+
+    Remove a site from the list of organizational news sites, without prompting
+    for confirmation
+      ${chalk.grey(config.delimiter)} ${commands.ORGNEWSSITE_REMOVE} --url https://contoso.sharepoint.com/sites/site1 --confirm
+  `);
+  }
+}
+
+module.exports = new SpoOrgNewsSiteRemoveCommand();